Refactored into nicer directory structure
